Add support for application calls
Move `bpf_get_call()` into appcode
