+++ bpf/include/bpf_appcalls.h
@@ -1,0 +1,2 @@
+// Default definition if header not defined by application
+#define BPF_SYSCALL_APP(XX)
