+++ src/call.cpp
@@ -7,25 +7,10 @@
  * directory for more details.
  */
 
-#include <cstdint>
-#include <cstdlib>
 #include <debug_progmem.h>
-
 #include <bpf.h>
 #include "bpf/store.h"
 #include "bpf/call.h"
-
-bpf_call_t bpf_get_call(uint32_t num)
-{
-	switch(num) {
-#define XX(function_code, function_name, return_type, ...)                                                             \
-	case function_code:                                                                                                \
-		return reinterpret_cast<bpf_call_t>(rBPF::VM::function_name);
-		BPF_SYSCALL_MAP(XX)
-	default:
-		return nullptr;
-	}
-}
 
 namespace rBPF
 {
